[Migration 4-3] Feature rollup/zk-sync
