+++ client/features/rollup/zk-sync/mocks/txn-batch.ts
@@ -1,4 +1,4 @@
-import type { ZkSyncBatch } from 'types/api/zkSyncL2';
+import type { ZkSyncBatch } from 'client/features/rollup/zk-sync/types/api';
 
 export const base: ZkSyncBatch = {
   commit_transaction_hash: '0x7cd80c88977c2b310f79196b0b2136da18012be015ce80d0d9e9fe6cfad52b16',
@@ -12,7 +12,7 @@ export const base: ZkSyncBatch = {
   l2_transactions_count: 287,
   number: 8051,
   prove_transaction_hash: '0xb424162ba5afe17c710dceb5fc8d15d7d46a66223454dae8c74aa39f6802625b',
-  prove_transaction_timestamp: '2022-03-19T10:29:05.279179Z',
+  prove_transaction_timestamp: '2022-03-19T10:29:05.279177Z',
   root_hash: '0x108c635b94f941fcabcb85500daec2f6be4f0747dff649b1cdd9dd7a7a264792',
   start_block_number: 1245209,
   status: 'Executed on L1',
