+++ client/features/rollup/zk-sync/mocks/txn-batches.ts
@@ -1,4 +1,4 @@
-import type { ZkSyncBatchesItem, ZkSyncBatchesResponse } from 'types/api/zkSyncL2';
+import type { ZkSyncBatchesItem, ZkSyncBatchesResponse } from 'client/features/rollup/zk-sync/types/api';
 
 export const sealed: ZkSyncBatchesItem = {
   commit_transaction_hash: null,
@@ -28,12 +28,12 @@ export const sent: ZkSyncBatchesItem = {
 
 export const executed: ZkSyncBatchesItem = {
   commit_transaction_hash: '0xa2628f477e1027ac1c60fa75c186b914647769ac1cb9c7e1cab50b13506a0035',
-  commit_transaction_timestamp: '2022-03-19T11:52:18.963659Z',
+  commit_transaction_timestamp: '2022-03-19T11:52:18.963070Z',
   execute_transaction_hash: '0xb7bd6b2b17498c66d3f6e31ac3685133a81b7f728d4f6a6f42741daa257d0d68',
-  execute_transaction_timestamp: '2022-03-19T13:28:16.712656Z',
+  execute_transaction_timestamp: '2022-03-19T13:28:16.712706Z',
   number: 8053,
   prove_transaction_hash: '0x9d44f2b775bd771f8a53205755b3897929aa672d2cd419b3b988c16d41d4f21e',
-  prove_transaction_timestamp: '2022-03-19T13:28:16.603104Z',
+  prove_transaction_timestamp: '2022-03-19T13:28:16.603046Z',
   status: 'Executed on L1',
   timestamp: '2022-03-19T10:01:52.000000Z',
   transactions_count: 1071,
