+++ client/features/rollup/zk-sync/pages/batch-details/ZkSyncL2TxnBatch.tsx
@@ -10,19 +10,21 @@ import useApiQuery from 'client/api/hooks/useApiQuery';
 import TxsWithFrontendSorting from 'client/slices/tx/pages/index/list/TxsWithFrontendSorting';
 import { TX } from 'client/slices/tx/stubs/tx';
 
+import { ZKSYNC_L2_TXN_BATCH } from 'client/features/rollup/zk-sync/stubs';
+
 import throwOnAbsentParamError from 'client/shared/errors/throw-on-absent-param-error';
 import throwOnResourceLoadError from 'client/shared/errors/throw-on-resource-load-error';
 import useIsMobile from 'client/shared/hooks/useIsMobile';
 import getQueryParamString from 'client/shared/router/get-query-param-string';
 
 import { generateListStub } from 'stubs/utils';
-import { ZKSYNC_L2_TXN_BATCH } from 'stubs/zkSyncL2';
 import RoutedTabs from 'toolkit/components/RoutedTabs/RoutedTabs';
 import TextAd from 'ui/shared/ad/TextAd';
 import PageTitle from 'ui/shared/Page/PageTitle';
 import Pagination from 'ui/shared/pagination/Pagination';
 import useQueryWithPages from 'ui/shared/pagination/useQueryWithPages';
-import ZkSyncL2TxnBatchDetails from 'ui/txnBatches/zkSyncL2/ZkSyncL2TxnBatchDetails';
+
+import ZkSyncL2TxnBatchDetails from './ZkSyncL2TxnBatchDetails';
 
 const TAB_LIST_PROPS = {
   marginBottom: 0,
