+++ client/features/rollup/zk-sync/pages/batch-details/__screenshots__/ZkSyncL2TxnBatch.pw.tsx_default_mobile-base-view-1.png
