+++ client/features/rollup/zk-sync/pages/batches/__screenshots__/ZkSyncL2TxnBatches.pw.tsx_mobile_base-view-mobile-1.png
