+++ client/features/rollup/zk-sync/utils/format-txn-batch-status.ts
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: LicenseRef-Blockscout
+
+import type { ZKSYNC_L2_TX_BATCH_STATUSES } from 'client/features/rollup/zk-sync/types/api';
+
+import { layerLabels } from 'lib/rollups/utils';
+
+export const formatZkSyncL2TxnBatchStatus = (status: typeof ZKSYNC_L2_TX_BATCH_STATUSES[number]) => {
+  return status.replace('L2', layerLabels.current).replace('L1', layerLabels.parent);
+};
