+++ pages/batches/index.tsx
@@ -16,7 +16,7 @@ const Batches = dynamic(() => {
 
   switch (rollupFeature.type) {
     case 'zkSync':
-      return import('ui/pages/ZkSyncL2TxnBatches');
+      return import('client/features/rollup/zk-sync/pages/batches/ZkSyncL2TxnBatches');
     case 'optimistic':
       return import('client/features/rollup/optimism/pages/batches/OptimisticL2TxnBatches');
     case 'arbitrum':
